feat: add /profile and make userContext
chore: add folder to component
