+++ frontend/src/api/authApi.ts
@@ -11,3 +11,16 @@ export const registerUser = async (data: RegisterData) => {
 export const loginUser = async (data: LoginData) => {
   return await api.post("/api/auth/login", data);
 };
+
+export const getProfile = async () => {
+  // Gọi API thật ở đây. Dưới đây là data mock:
+  
+  return {
+    id: "1",
+    username: "nguyenvana",
+    email: "van.a@example.com",
+    avatarUrl: "",
+    bio: "Sinh viên CNTT năm 3.",
+    createdAt: "2024-02-12T08:00:00Z",
+  };
+};
