+++ frontend/src/components/auth/AuthHeader.css
@@ -1,0 +1,16 @@
+.header-btn {
+    height: max-content !important;
+    padding: 0.5rem 1.5rem !important;
+    border-radius: 5px !important;
+    background-color: #FDC886 !important;
+    color: #2d2d2d !important;
+    border: none !important;
+    font-weight: 600 !important;
+}
+
+h1 {
+    font-size: 2rem !important;
+    font-weight: 900 !important;
+    color: #000 !important;
+    margin: 0 !important;
+}
