+++ frontend/src/components/auth/RegisterHeader.tsx
