+++ frontend/src/components/profile/UserProfile.css
@@ -1,0 +1,20 @@
+.cover-photo {
+    background: url("https://media-hosting.imagekit.io/b033b794cbba4b35/490191769_1077625737737226_5317678520741036604_n.jpg?Expires=1840579515&Key-Pair-Id=K2ZIVPTIP2VGHC&Signature=zGKm8brxNl4XSyHUIvKbHlCYExx3o6ppffiYnHB1cl4muP7AQTsbGeSNBq21VpeO7g3RC7FsXFxwdjAJec3XKQ5hVghjgEhvrImCvNVjrqzYWmotDDdm7C1g7V3nNVJ-uT2sFe9x1z-XqTiaKYGqWHz7Vr8wqUgcylzh-Xa0924m9L0~Xk3eciVbQKrrfwYaFiih9VTdOsigOPjF9j14A2kAefQ11859edcOkbSx3Lge1~-hcfpP5r3s4McbSh1VwK8W-IMpA8CoRVeOCxL~X8vGJ~cph6e1ZQMvvy3VkzC6LC5uHKkfI4DDJ2~J919-Kg3P2e169JwUfX16DBRRBQ__") center/cover;
+    height: 150px;
+  }
+
+  .avatar {
+    width: 150px;
+    height: 150px;
+    object-fit: cover;
+    border-radius: 50%;
+    border: 4px solid white;
+    margin-top: -75px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  }
+  
+  .profile-header {
+    margin-top: -60px;
+    padding: 20px 0;
+  }
+
