+++ frontend/src/interfaces/UserProfile.ts
@@ -1,0 +1,9 @@
+export interface UserProfile {
+    id: string;
+    username: string;
+    email: string;
+    picUrl?: string;
+    // bio?: string;
+    // createdAt: string;
+  }
+
