+++ frontend/src/main.tsx
@@ -3,9 +3,12 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "./index.css";
 import App from "./App.tsx";
 import { BrowserRouter } from "react-router-dom";
+import { UserProvider } from "./context/UserContext";
 
 createRoot(document.getElementById("root")!).render(
   <BrowserRouter>
-    <App />
+    <UserProvider>
+      <App />
+    </UserProvider>
   </BrowserRouter>
 );
