+++ frontend/src/routes/AppRoutes.tsx
@@ -3,6 +3,7 @@ import HomePage from "../pages/HomePage";
 import LoginPage from "../pages/LoginPage";
 import RegisterPage from "../pages/RegisterPage";
 import RegisterComplete from "../pages/RegisterCompletePage";
+import ProfilePage from "../pages/ProfilePage";
 
 export default function AppRoutes() {
   return (
@@ -16,6 +17,7 @@ export default function AppRoutes() {
       <Route path="/login" element={<LoginPage />} />
       <Route path="/register" element={<RegisterPage />} />
       <Route path="/register-complete" element={<RegisterComplete />} />
+      <Route path="/profile" element={<ProfilePage />} />
     </Routes>
   );
 }
